resurrect old declaration of getfp (s/int/size_t/ arg)

+++ mon.h
@@ -95,7 +95,7 @@ void		 nc_runall(void (*)(struct fill *));
 void		 nc_runsn(void (*)(struct fill *));
 void		 nc_apply(void (*)(struct fill *), size_t);
 void		 nc_set(int);
-struct fill	*nc_getfp(int);
+struct fill	*nc_getfp(size_t);
 
 /* png.c */
 void 		*png_load(char *, unsigned int *, unsigned int *);
